account: add security Q&A to recover password
Using just the email to reset password is way too easy.
Make it harder by storing personal security Q&A.

Security question is stored the same way as the email,
    since the program needs to decrypt and show upon recovery.

Security answer is stored the same way as the password,
    to make the program itself unaware of the original answer.

Now that additional step is added to the recovery process,
allow admin account to be recovered as well.